+++ src/main/java/com/martak/adventofcode/intcodecomputer/model/argmode/ArgMode.java
@@ -2,8 +2,10 @@ package com.martak.adventofcode.intcodecomputer.model.argmode;
 
 import com.martak.adventofcode.intcodecomputer.model.Context;
 
+import java.math.BigInteger;
+
 public interface ArgMode {
 
-    int getValue(Context context, int pointer);
+    BigInteger getValue(Context context, int pointer);
 
 }
