+++ src/main/java/com/martak/adventofcode/intcodecomputer/model/argmode/ImmediateArgMode.java
@@ -2,11 +2,13 @@ package com.martak.adventofcode.intcodecomputer.model.argmode;
 
 import com.martak.adventofcode.intcodecomputer.model.Context;
 
+import java.math.BigInteger;
+
 public class ImmediateArgMode implements ArgMode {
 
     @Override
-    public int getValue(Context context, int pointer) {
-        int[] code = context.getCode();
+    public BigInteger getValue(Context context, int pointer) {
+        BigInteger[] code = context.getCode();
         return code[pointer];
     }
 }
