+++ src/main/java/com/martak/adventofcode/intcodecomputer/model/argmode/PositionArgMode.java
@@ -2,12 +2,14 @@ package com.martak.adventofcode.intcodecomputer.model.argmode;
 
 import com.martak.adventofcode.intcodecomputer.model.Context;
 
+import java.math.BigInteger;
+
 public class PositionArgMode implements ArgMode {
 
     @Override
-    public int getValue(Context context, int pointer) {
-        int[] code = context.getCode();
-        int address = code[pointer];
+    public BigInteger getValue(Context context, int pointer) {
+        BigInteger[] code = context.getCode();
+        int address = code[pointer].intValue();
         return code[address];
     }
 }
